Fix an int overflow issue in the interpolation search on big arrays

+++ src/main/java/net/adeptropolis/frogspawn/graphs/implementations/arrays/InterpolationSearch.java
@@ -37,7 +37,7 @@ public class InterpolationSearch {
 
     while (key >= lowVal && key <= highVal) {
 
-      long mid = (int) (low + (((key - lowVal) * (high - low)) / (highVal - lowVal)));
+      long mid = low + (key - lowVal) * (high - low) / (highVal - lowVal);
       int midVal = ints.get(mid);
 
       if (key > midVal) {
